refactor: 활동 시간, 모임 운동 타입, 모임 정렬 타입 Enum 모임 도메인으로 변경
