+++ src/main/java/umc/cockple/demo/domain/contest/dto/ContestRecordCreateDTO.java
@@ -8,7 +8,7 @@ import lombok.Builder;
 import org.springframework.web.multipart.MultipartFile;
 import umc.cockple.demo.global.enums.Level;
 import umc.cockple.demo.domain.contest.enums.MedalType;
-import umc.cockple.demo.global.enums.ParticipationType;
+import umc.cockple.demo.domain.party.enums.ParticipationType;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
