+++ src/main/java/umc/cockple/demo/domain/contest/dto/ContestRecordDetailDTO.java
@@ -3,7 +3,7 @@ package umc.cockple.demo.domain.contest.dto;
 import lombok.Builder;
 import umc.cockple.demo.global.enums.Level;
 import umc.cockple.demo.domain.contest.enums.MedalType;
-import umc.cockple.demo.global.enums.ParticipationType;
+import umc.cockple.demo.domain.party.enums.ParticipationType;
 
 import java.time.LocalDate;
 import java.util.List;
