+++ src/main/java/umc/cockple/demo/domain/contest/dto/ContestRecordSimpleDTO.java
@@ -2,7 +2,7 @@ package umc.cockple.demo.domain.contest.dto;
 
 import lombok.Builder;
 import umc.cockple.demo.global.enums.Level;
-import umc.cockple.demo.global.enums.ParticipationType;
+import umc.cockple.demo.domain.party.enums.ParticipationType;
 
 import java.time.LocalDate;
