+++ src/main/java/umc/cockple/demo/domain/party/enums/ActivityTime.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.party.enums;
 
 import lombok.Getter;
