+++ src/main/java/umc/cockple/demo/domain/party/enums/PartyOrderType.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.global.enums;
+package umc.cockple.demo.domain.party.enums;
 
 import lombok.Getter;
 import umc.cockple.demo.domain.party.exception.PartyErrorCode;
